Revert "Merge pull request #2 from LLercpp/Encapsulation"
This reverts commit ffeb350ec16fb6f4071a6ce6ed40c76301585290, reversing
changes made to 29fccffa2b6ecb8d73d4a9683a02e9da9ab7a075.
