+++ src/seedu/addressbook/data/person/Address.java
@@ -6,12 +6,14 @@ import seedu.addressbook.data.exception.IllegalValueException;
  * Represents a Person's address in the address book.
  * Guarantees: immutable; is valid as declared in {@link #isValidAddress(String)}
  */
-public class Address extends Contact{
+public class Address {
 
     public static final String EXAMPLE = "123, some street";
     public static final String MESSAGE_ADDRESS_CONSTRAINTS = "Person addresses can be in any format";
     public static final String ADDRESS_VALIDATION_REGEX = ".+";
 
+    public final String value;
+    private boolean isPrivate;
 
     /**
      * Validates given address.
@@ -19,11 +21,12 @@ public class Address extends Contact{
      * @throws IllegalValueException if given address string is invalid.
      */
     public Address(String address, boolean isPrivate) throws IllegalValueException {
-        super (address, isPrivate);
         String trimmedAddress = address.trim();
+        this.isPrivate = isPrivate;
         if (!isValidAddress(trimmedAddress)) {
             throw new IllegalValueException(MESSAGE_ADDRESS_CONSTRAINTS);
         }
+        this.value = trimmedAddress;
     }
 
     /**
@@ -33,5 +36,24 @@ public class Address extends Contact{
         return test.matches(ADDRESS_VALIDATION_REGEX);
     }
 
+    @Override
+    public String toString() {
+        return value;
+    }
 
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Address // instanceof handles nulls
+                && this.value.equals(((Address) other).value)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
 }
