+++ src/seedu/addressbook/ui/TextUi.java
@@ -129,7 +129,7 @@ public class TextUi {
         if (resultPersons.isPresent()) {
             showPersonListView(resultPersons.get());
         }
-        showToUser(result.getFeedbackToUser(), DIVIDER);
+        showToUser(result.feedbackToUser, DIVIDER);
     }
 
     /**
